Define suckless fonts using "pixelsize" on UHD displays
Although it is possible to configure dwm and dmenu to have the same bar
height using both "size" and "pixelsize", these properties produce
fights that have slightly different kerning, line and glyph height
likely due to the impact of DPI settings affecting font specs using
"size". This change makes it so "pixelsize" is used for both UHD and
non-UHD displays to maintain similar styling of the text.