Get Cmd+(Shift)+Tab working as intended
Due to some changes in macOS, Karabiner and/or iTerm2, Cmd+Shift+Tab
could no longer be mapped to Ctrl+Shift+Tab only within iTerm2. With the
original configuration that worked on some earlier release of macOS,
pressing Cmd+(Shift)+Tab would bring up the application switcher. To
resolve this, Cmd+(Shift)+Tab in iTerm2 is now mapped to the same
terminal escape sequences used for Ctrl+(Shift)+Tab. Accompanying this
are changes to the Karabiner Elements configuration that maps
Cmd+(Shift)+Tab to Ctrl+(Shift)+Tab, and a refactor of how the
configuration files are handled to capture all of Karabiner Element's
configuration state.