Add Windows Terminal configuration
This adds the Windows Terminal JSON configuration file. For now, it will
not be managed by the usual configuration system because I have not
decided how to handle the fact that I cannot update "settings.json"
while Windows Terminal is running and that I cannot have a symlink from
Windows pointing into the WSL environment. There is a target for copying
the Windows Terminal configuration out of the host Windows environment
that uses "jq" to normalize the JSON to mitigate no-op reordering and
other formatting quirks like
<https://github.com/microsoft/terminal/issues/10887> associated with the
terminal's use of jsoncpp.